+++ src/WebJobs.Extensions.ApiHub/ApiHubFileTriggerAttribute.cs
@@ -9,7 +9,7 @@ namespace Microsoft.Azure.WebJobs
     public sealed class ApiHubFileTriggerAttribute : ApiHubFileAttribute
     {
         /// <summary>
-        /// Create an instance of an attribute that points to the file on SAAS file trigger
+        /// Attribute used to bind a parameter to a SAAS file
         /// </summary>
         /// <param name="key">App settings key name that have the connections string</param>
         /// <param name="path">Relative path to the file <example>/folder/subfolder/file.txt</example></param>
